Added/Updated Daos, Object Classes, DaoConfig, and Dao Keeper (including beans)

+++ src/main/java/edu/ben/template/controller/UserController.java
@@ -19,7 +19,7 @@ public class UserController extends BaseController {
 	public String displayAccount(Model model) {
 		long userId = getCurrentUserId();
 
-		User user = getUserDao().getObjectById(userId);
+		User user = getUserDao().getObjectById((int) userId);
 
 		model.addAttribute("user", user);
 		return "account/view";
